FEATURE - adicionado os dados comparados aos benchmarks * Criados novas classes model, dto, service e repository * Criado três migrations para adição e alteração de dados * Ajustado retorno da api de benchmark

+++ src/main/java/br/com/covidbenchmarkapi/domain/dto/CovidUfDto.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 public class CovidUfDto {
     private String data;
     private String uf;
-    private int confirmacoes;
-    private int mortes;
+    private long confirmacoes;
+    private long mortes;
     private float confirmadosSobreCemMilHabitantes;
     private float mortesSobreConfirmados;
     private boolean possuiDados;
